requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Logging
python-json-logger==2.0.7

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Environment variables
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
ruff==0.1.14
mypy==1.8.0

# Optional: For future Azure AD integration
# msal==1.26.0
# azure-identity==1.15.0

# Optional: For future AWS integration
# boto3==1.34.34

# Optional: For future observability
# opentelemetry-api==1.22.0
# opentelemetry-sdk==1.22.0
# opentelemetry-instrumentation-fastapi==0.43b0